pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "golden-path-strands"
version = "0.1.0"
description = "Golden Path Strands framework"
readme = "README.md"
requires-python = ">=3.8"
authors = [{name = "John Boren"}]
dependencies = [
    "pyyaml",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "isort",
    "pre-commit",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
